Test 1: I declare a variable and assigned a value, 0. I initialized the condition and compared "i" to numbers of the values in the array where is less than the length I completed the for loop syntax by returning the "sum" of the values of numbers the array.
Test 2: I declare and assigned figure 0 to the variable, count
I initialized the "for loop" syntax and compared i to the number length of the values in the array (where i < arrayOfNumbers.length).
I selected out the number divisible by 2 (without remainder) with code arrayOfNumbers[i] % 2 == 0 whwre i increases by one till the condition in comparison is met.
I completed the "for loop" syntax with the "return" of the variable "count".